Recurse refetch
